pyproject.toml
[project]
name = "railway-fastapi-template"
version = "0.1.0"
description = "This example starts up a FastAPI server"
authors = [
    {name = "DeviousLab", email = "deviouslab@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.18.3",
    "langchain>=1.2.0",
    "openai>=1.6.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.0",
    "discord-py>=2.2.2",
    "rich>=13.3.4",
    "requests>=2.29.0",
    "beautifulsoup4>=4.12.2",
    "requests-oauthlib>=1.3.1",
    "tweepy>=4.14.0",
    "google-genai>=0.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.116.0",
    "google-auth-oauthlib>=1.2.0",
    "tavily>=1.1.0",
    "langgraph>=1.0.5",
    "langchain-openai>=1.1.6",
]

[tool.poetry]
name = "railway-fastapi-template"
version = "0.1.0"
description = "This example starts up a FastAPI server"
authors = ["DeviousLab <deviouslab@gmail.com>"]
license = "MIT"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.18.3"
langchain = "^1.2.0"
openai = "^1.6.1"
python-decouple = "^3.8"
python-dotenv = "^1.0.0"
discord-py = "^2.2.2"
rich = "^13.3.4"
requests = "^2.29.0"
beautifulsoup4 = "^4.12.2"
requests-oauthlib = "^1.3.1"
tweepy = "^4.14.0"
google-genai = "^0.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.116.0"
google-auth-oauthlib = "^1.2.0"

[tool.poetry.group.dev.dependencies]
